Player interaction, and WARNING
The player can interact with the mirrors now and turn those. To make this work I used the player controller and third person movement script. This also means that potential conflicts can arise. These shouldn't be too bad, the portions of code I wrote can just be added to the bottom of the script during the merge and it should work.